# docs/swagger.yaml
definitions:
...
  exception.AppErr:
    properties:
      message:
        example: some error
        type: string
      status:
        example: 404
        type: integer
    type: object
  exception.ValidationError:
    properties:
      errors:
        items:
          $ref: '#/definitions/exception.validationErrorItem'
        type: array
      status:
        example: 400
        type: integer
    type: object
  exception.validationErrorItem:
    properties:
      key:
        example: email
        type: string
      message:
        example: email is invalid
        type: string
    type: object
...
  role.AddRoleToUserDto:
    properties:
      title:
        example: ADMIN
        minLength: 3
        type: string
      user_id:
        example: 1
        minimum: 1
        type: integer
    required:
    - title
    - user_id
    type: object
  role.CreateRoleDto:
    properties:
      title:
        example: ADMIN
        maxLength: 55
        minLength: 6
        type: string
    required:
    - title
    type: object
  role.Role:
    properties:
      role_id:
        example: 1
        type: integer
      title:
        example: ADMIN
        type: string
    type: object
  user.ChangePasswordDto:
    properties:
      code:
        example: "123456"
        maxLength: 6
        minLength: 6
        type: string
      new_password:
        example: sdfsdfs222
        maxLength: 100
        minLength: 6
        type: string
      old_password:
        example: sdfsdfs222
        maxLength: 100
        minLength: 6
        type: string
    required:
    - code
    - new_password
    - old_password
    type: object
...
  user.User:
    properties:
      email:
        example: makc@mail.ru
        type: string
      is_activated:
        example: false
        type: boolean
      roles:
        items:
          $ref: '#/definitions/role.Role'
        type: array
      user_id:
        example: 1
        type: integer
    required:
    - email
    type: object
info:
  contact: {}
paths:
...
  /api/role/:
    post:
      consumes:
      - application/json
      description: Create role by body arguments
      parameters:
      - description: create role with body dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/role.CreateRoleDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/role.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      security:
      - BearerToken: []
      summary: Create role
      tags:
      - roles
  /api/role/add-to-user:
    post:
      consumes:
      - application/json
      description: Add role to user by body arguments
      parameters:
      - description: add role to user with body dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/role.AddRoleToUserDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      security:
      - BearerToken: []
      summary: Add role to user
      tags:
      - roles
  /api/role/remove-from-user:
    delete:
      consumes:
      - application/json
      description: Remove role from user by body arguments
      parameters:
      - description: Remove role from user with body dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/role.AddRoleToUserDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      security:
      - BearerToken: []
      summary: Remove role from user
      tags:
      - roles
  /api/user/by-id/:id:
    get:
      consumes:
      - application/json
      description: Get user by id
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Get user by id
      tags:
      - users
  /api/user/change-password:
    patch:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: change password
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/user.ChangePasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/exception.AppErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      security:
      - BearerToken: []
      summary: Change password
      tags:
      - users
  /api/user/create-change-password-code:
    post:
      consumes:
      - application/json
      description: Create change password code and send to email
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/exception.AppErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      security:
      - BearerToken: []
      summary: Create change password code
      tags:
      - users
  /api/user/lk:
    get:
      consumes:
      - application/json
      description: Get profile info by auth only
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/exception.AppErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      security:
      - BearerToken: []
      summary: Get profile info
      tags:
      - users
...
swagger: "2.0"